requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.4
pandas~=2.2.3
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.41
python-dotenv~=1.1.0